feat(app): preload and cache game library for instant open
- perf(vdf_parser): refactor parse_remotecache to single-pass state machine; merge timestamp handling; add extract_key_value
- perf(app): reuse VdfParser in update_local_save_paths to avoid per-file init
- ux(app): game selector shows scanning/empty states; avoid auto-opening on scan completion
- fix(app): remove unused size format helpers; resolve clippy warnings
- fix(steam_api): align CloudFile.size to u64 and use raw size from API

+++ src/steam_api.rs
@@ -15,7 +15,7 @@ pub struct SteamCloudManager {
 #[derive(Debug, Clone)]
 pub struct CloudFile {
     pub name: String,
-    pub size: i32,
+    pub size: u64,
     pub timestamp: DateTime<Local>,
     pub is_persisted: bool,
     pub exists: bool,
@@ -220,7 +220,7 @@ impl SteamCloudManager {
 
             let file = CloudFile {
                 name: steam_file.name.clone(),
-                size: steam_file.size as i32,
+                size: steam_file.size,
                 timestamp,
                 is_persisted: steam_file_handle.is_persisted(),
                 exists: steam_file_handle.exists(),
